+++ attribute/kv.go
@@ -36,14 +36,29 @@ func Bool(k string, v bool) KeyValue {
 	return Key(k).Bool(v)
 }
 
+// BoolSlice creates a KeyValue with a BOOLSLICE Value type.
+func BoolSlice(k string, v []bool) KeyValue {
+	return Key(k).BoolSlice(v)
+}
+
 // Int creates a KeyValue with an INT64 Value type.
 func Int(k string, v int) KeyValue {
 	return Key(k).Int(v)
 }
 
-// Int64 creates a KeyValue with a INT64 Value type.
+// IntSlice creates a KeyValue with an INT64SLICE Value type.
+func IntSlice(k string, v []int) KeyValue {
+	return Key(k).IntSlice(v)
+}
+
+// Int64 creates a KeyValue with an INT64 Value type.
 func Int64(k string, v int64) KeyValue {
 	return Key(k).Int64(v)
+}
+
+// Int64Slice creates a KeyValue with an INT64SLICE Value type.
+func Int64Slice(k string, v []int64) KeyValue {
+	return Key(k).Int64Slice(v)
 }
 
 // Float64 creates a KeyValue with a FLOAT64 Value type.
@@ -51,9 +66,19 @@ func Float64(k string, v float64) KeyValue {
 	return Key(k).Float64(v)
 }
 
+// Float64Slice creates a KeyValue with a FLOAT64SLICE Value type.
+func Float64Slice(k string, v []float64) KeyValue {
+	return Key(k).Float64Slice(v)
+}
+
 // String creates a KeyValue with a STRING Value type.
 func String(k, v string) KeyValue {
 	return Key(k).String(v)
+}
+
+// StringSlice creates a KeyValue with a STRINGSLICE Value type.
+func StringSlice(k string, v []string) KeyValue {
+	return Key(k).StringSlice(v)
 }
 
 // Stringer creates a new key-value pair with a passed name and a string
@@ -64,6 +89,8 @@ func Stringer(k string, v fmt.Stringer) KeyValue {
 
 // Array creates a new key-value pair with a passed name and a array.
 // Only arrays of primitive type are supported.
+//
+// Deprecated: Use the typed *Slice functions instead.
 func Array(k string, v interface{}) KeyValue {
 	return Key(k).Array(v)
 }
@@ -82,8 +109,24 @@ func Any(k string, value interface{}) KeyValue {
 	rv := reflect.ValueOf(value)
 
 	switch rv.Kind() {
-	case reflect.Array, reflect.Slice:
-		return Array(k, value)
+	case reflect.Array:
+		rv = rv.Slice(0, rv.Len())
+		fallthrough
+	case reflect.Slice:
+		switch reflect.TypeOf(value).Elem().Kind() {
+		case reflect.Bool:
+			return BoolSlice(k, rv.Interface().([]bool))
+		case reflect.Int:
+			return IntSlice(k, rv.Interface().([]int))
+		case reflect.Int64:
+			return Int64Slice(k, rv.Interface().([]int64))
+		case reflect.Float64:
+			return Float64Slice(k, rv.Interface().([]float64))
+		case reflect.String:
+			return StringSlice(k, rv.Interface().([]string))
+		default:
+			return KeyValue{Key: Key(k), Value: Value{vtype: INVALID}}
+		}
 	case reflect.Bool:
 		return Bool(k, rv.Bool())
 	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64: