+++ exporters/zipkin/model.go
@@ -181,9 +181,18 @@ func toZipkinTags(data tracesdk.ReadOnlySpan) map[string]string {
 	m := make(map[string]string, len(attr)+len(extraZipkinTags))
 	for _, kv := range attr {
 		switch kv.Value.Type() {
-		// For array attributes, serialize as JSON list string.
-		case attribute.ARRAY:
-			json, _ := json.Marshal(kv.Value.AsArray())
+		// For slice attributes, serialize as JSON list string.
+		case attribute.BOOLSLICE:
+			json, _ := json.Marshal(kv.Value.AsBoolSlice())
+			m[(string)(kv.Key)] = (string)(json)
+		case attribute.INT64SLICE:
+			json, _ := json.Marshal(kv.Value.AsInt64Slice())
+			m[(string)(kv.Key)] = (string)(json)
+		case attribute.FLOAT64SLICE:
+			json, _ := json.Marshal(kv.Value.AsFloat64Slice())
+			m[(string)(kv.Key)] = (string)(json)
+		case attribute.STRINGSLICE:
+			json, _ := json.Marshal(kv.Value.AsStringSlice())
 			m[(string)(kv.Key)] = (string)(json)
 		default:
 			m[(string)(kv.Key)] = kv.Value.Emit()